# Vectorize block_diag to avoid inner loops over block elements