--- a/benchmarks/benchmarks/sparse.py
+++ b/benchmarks/benchmarks/sparse.py
@@ -192,10 +192,10 @@ class Construction(Benchmark):
             T[i, j] = v
 
 
-class BlockDiagConstruction(Benchmark):
+class BlockDiagDenseConstruction(Benchmark):
     param_names = ['num_matrices']
     params = [1000, 5000, 10000, 15000, 20000]
-    
+
     def setup(self, num_matrices):
         self.matrices = []
         for i in range(num_matrices):
@@ -203,9 +203,25 @@ class BlockDiagConstruction(Benchmark):
             columns = np.random.randint(1, 4)
             mat = np.random.randint(0, 10, (rows, columns))
             self.matrices.append(mat)
-            
+
     def time_block_diag(self, num_matrices):
-        sparse.block_diag(self.matrices)            
+        sparse.block_diag(self.matrices)
+
+
+class BlockDiagSparseConstruction(Benchmark):
+    param_names = ['num_matrices']
+    params = [100, 500, 1000, 1500, 2000]
+
+    def setup(self, num_matrices):
+        self.matrices = []
+        for i in range(num_matrices):
+            rows = np.random.randint(1, 20)
+            columns = np.random.randint(1, 20)
+            mat = np.random.randint(0, 10, (rows, columns))
+            self.matrices.append(mat)
+
+    def time_block_diag(self, num_matrices):
+        sparse.block_diag(self.matrices)
 
 
 class Conversion(Benchmark):