This PR modifies block_diag to avoid inner loops over the block entries via vectorization and early conversions to COO.  This results in comparable performance with small dense blocks, but better performance with large dense blocks or sparse blocks.

Before (commit 90cfcbd1291930929736b7c8d182168a456bd8d6):

```
[100.00%] ··· sparse.BlockDiagDenseConstruction.time_block_diag                                       ok
[100.00%] ··· ============== ============
               num_matrices              
              -------------- ------------
                   1000       8.01±0.3ms 
                   5000       37.3±0.9ms 
                  10000        79.3±1ms  
                  15000        122±4ms   
                  20000        164±2ms   
              ============== ============

[100.00%] ··· sparse.BlockDiagSparseConstruction.time_block_diag                                      ok
[100.00%] ··· ============== ==========
               num_matrices            
              -------------- ----------
                   100        13.5±1ms 
                   500        69.5±1ms 
                   1000       150±3ms  
                   1500       221±7ms  
                   2000       296±4ms  
              ============== ==========
```

After (commit bfa3269d108be255aab7420f0be20ee5662155a1):

```
[100.00%] ··· sparse.BlockDiagDenseConstruction.time_block_diag                                       ok
[100.00%] ··· ============== ============
               num_matrices              
              -------------- ------------
                   1000       7.57±0.1ms 
                   5000        38.9±1ms  
                  10000       75.1±0.9ms 
                  15000        115±2ms   
                  20000        158±8ms   
              ============== ============

[100.00%] ··· sparse.BlockDiagSparseConstruction.time_block_diag                                      ok
[100.00%] ··· ============== ============
               num_matrices              
              -------------- ------------
                   100         995±30μs  
                   500        4.98±0.1ms 
                   1000       10.2±0.4ms 
                   1500       15.6±0.9ms 
                   2000       21.2±0.6ms 
              ============== ============
```